52 Badgers Dene
But the inspector was mistaken, for those criminals were not destined to fall into the hands of
justice. As we rolled into Eyford Station we saw a gigantic column of smoke which streamed up from
behind a small clump of trees in the neighbourhood and hung like an immense ostrich feather over the
landscape.

"A house on fire?" asked Bradstreet as the train steamed off again on its way.

"Yes, sir!" said the station-master.

"When did it break out?"