38 Balcombe Close
The name was indeed well known to us as belonging to the senior partner in the second largest
private banking concern in the City of London. What could have happened, then, to bring one of the
foremost citizens of London to this most pitiable pass? We waited, all curiosity, until with another
effort he braced himself to tell his story.

"I feel that time is of value," said he; "that is why I hastened here when the police inspector
suggested that I should secure your co-operation. I came to Baker Street by the Underground and
hurried from there on foot, for the cabs go slowly through this snow. That is why I was so out of
breath, for I am a man who takes very little exercise. I feel better now, and I will put the facts
before you as shortly and yet as clearly as I can.