1 Corbet Close
"I was shaken but not hurt by the fall; so I picked myself up and rushed off among the bushes as
hard as I could run, for I understood that I was far from being out of danger yet. Suddenly,
however, as I ran, a deadly dizziness and sickness came over me. I glanced down at my hand, which
was throbbing painfully, and then, for the first time, saw that my thumb had been cut off and that
the blood was pouring from my wound. I endeavoured to tie my handkerchief round it, but there came a
sudden buzzing in my ears, and next moment I fell in a dead faint among the rose-bushes.

"How long I remained unconscious I cannot tell. It must have been a very long time, for the moon had
sunk, and a bright morning was breaking when I came to myself. My clothes were all sodden with dew,
and my coat-sleeve was drenched with blood from my wounded thumb. The smarting of it recalled in an
instant all the particulars of my night's adventure, and I sprang to my feet with the feeling that I
might hardly yet be safe from my pursuers. But to my astonishment, when I came to look round me,
neither house nor garden were to be seen. I had been lying in an angle of the hedge close by the
highroad, and just a little lower down was a long building, which proved, upon my approaching it, to
be the very station at which I had arrived upon the previous night. Were it not for the ugly wound
upon my hand, all that had passed during those dreadful hours might have been an evil dream.

"Half dazed, I went into the station and asked about the morning train. There would be one to
Reading in less than an hour. The same porter was on duty, I found, as had been there when I
arrived. I inquired of him whether he had ever heard of Colonel Lysander Stark. The name was strange
to him. Had he observed a carriage the night before waiting for me? No, he had not. Was there a
police-station anywhere near? There was one about three miles off.

"It was too far for me to go, weak and ill as I was. I determined to wait until I got back to town
before telling my story to the police. It was a little past six when I arrived, so I went first to
have my wound dressed, and then the doctor was kind enough to bring me along here. I put the case
into your hands and shall do exactly what you advise."