27 Beaconside
"But what is the good of all these vague theories," cried the banker impatiently, "when I have told
you that I saw Arthur with the coronet in his hands?"

"Wait a little, Mr. Holder. We must come back to that. About this girl, Miss Holder. You saw her
return by the kitchen door, I presume?"

"Yes; when I went to see if the door was fastened for the night I met her slipping in. I saw the
man, too, in the gloom."

"Do you know him?"

"Oh, yes! he is the green-grocer who brings our vegetables round.  His name is Francis Prosper."

"He stood," said Holmes, "to the left of the door--that is to say, farther up the path than is
necessary to reach the door?"