97 Wantage Close
"I was sure that I might rely on you."

"But what is it you wish?"

"When Mrs. Turner has brought in the tray I will make it clear to you. Now," he said as he turned
hungrily on the simple fare that our landlady had provided, "I must discuss it while I eat, for I
have not much time. It is nearly five now. In two hours we must be on the scene of action. Miss
Irene, or Madame, rather, returns from her drive at seven. We must be at Briony Lodge to meet her."

"And what then?"

"You must leave that to me. I have already arranged what is to occur. There is only one point on
which I must insist. You must not interfere, come what may. You understand?"

"I am to be neutral?"