50 Corbet Close
"Great heavens!" cried the inspector, "it is, indeed, the missing man. I know him from the
photograph."

The prisoner turned with the reckless air of a man who abandons himself to his destiny. "Be it so,"
said he. "And pray what am I charged with?"

"With making away with Mr. Neville St.--Oh, come, you can't be charged with that unless they make a
case of attempted suicide of it," said the inspector with a grin. "Well, I have been twenty-seven
years in the force, but this really takes the cake."