86 Holmefield Avenue
"You have erred, perhaps," he observed, taking up a glowing cinder with the tongs and lighting with
it the long cherry-wood pipe which was wont to replace his clay when he was in a disputatious rather
than a meditative mood--"you have erred perhaps in attempting to put colour and life into each of
your statements instead of confining yourself to the task of placing upon record that severe
reasoning from cause to effect which is really the only notable feature about the thing."

"It seems to me that I have done you full justice in the matter," I remarked with some coldness, for
I was repelled by the egotism which I had more than once observed to be a strong factor in my
friend's singular character.

"No, it is not selfishness or conceit," said he, answering, as was his wont, my thoughts rather than
my words. "If I claim full justice for my art, it is because it is an impersonal thing--a thing
beyond myself. Crime is common. Logic is rare. Therefore it is upon the logic rather than upon the
crime that you should dwell. You have degraded what should have been a course of lectures into a
series of tales."

It was a cold morning of the early spring, and we sat after breakfast on either side of a cheery
fire in the old room at Baker Street. A thick fog rolled down between the lines of dun-coloured
houses, and the opposing windows loomed like dark, shapeless blurs through the heavy yellow wreaths.
Our gas was lit and shone on the white cloth and glimmer of china and metal, for the table had not
been cleared yet. Sherlock Holmes had been silent all the morning, dipping continuously into the
advertisement columns of a succession of papers until at last, having apparently given up his
search, he had emerged in no very sweet temper to lecture me upon my literary shortcomings.

"At the same time," he remarked after a pause, during which he had sat puffing at his long pipe and
gazing down into the fire, "you can hardly be open to a charge of sensationalism, for out of these
cases which you have been so kind as to interest yourself in, a fair proportion do not treat of
crime, in its legal sense, at all. The small matter in which I endeavoured to help the King of
Bohemia, the singular experience of Miss Mary Sutherland, the problem connected with the man with
the twisted lip, and the incident of the noble bachelor, were all matters which are outside the pale
of the law. But in avoiding the sensational, I fear that you may have bordered on the trivial."

"The end may have been so," I answered, "but the methods I hold to have been novel and of interest."