164 Larch Walk
"Better close the front door," cried Holmes, and we all rushed down the stairs together. We had
hardly reached the hall when we heard the baying of a hound, and then a scream of agony, with a
horrible worrying sound which it was dreadful to listen to. An elderly man with a red face and
shaking limbs came staggering out at a side door.

"My God!" he cried. "Someone has loosed the dog. It's not been fed for two days. Quick, quick, or
it'll be too late!"

Holmes and I rushed out and round the angle of the house, with Toller hurrying behind us. There was
the huge famished brute, its black muzzle buried in Rucastle's throat, while he writhed and screamed
upon the ground. Running up, I blew its brains out, and it fell over with its keen white teeth still
meeting in the great creases of his neck. With much labour we separated them and carried him, living
but horribly mangled, into the house. We laid him upon the drawing-room sofa, and having dispatched
the sobered Toller to bear the news to his wife, I did what I could to relieve his pain. We were all
assembled round him when the door opened, and a tall, gaunt woman entered the room.