19 Corbet Close
Sherlock Holmes sat down beside him on the couch and patted him kindly on the shoulder.

"If you leave it to a court of law to clear the matter up," said he, "of course you can hardly avoid
publicity. On the other hand, if you convince the police authorities that there is no possible case
against you, I do not know that there is any reason that the details should find their way into the
papers. Inspector Bradstreet would, I am sure, make notes upon anything which you might tell us and
submit it to the proper authorities. The case would then never go into court at all."

"God bless you!" cried the prisoner passionately. "I would have endured imprisonment, ay, even
execution, rather than have left my miserable secret as a family blot to my children.

"You are the first who have ever heard my story. My father was a schoolmaster in Chesterfield, where
I received an excellent education. I travelled in my youth, took to the stage, and finally became a
reporter on an evening paper in London. One day my editor wished to have a series of articles upon
begging in the metropolis, and I volunteered to supply them. There was the point from which all my
adventures started. It was only by trying begging as an amateur that I could get the facts upon
which to base my articles. When an actor I had, of course, learned all the secrets of making up, and
had been famous in the green-room for my skill. I took advantage now of my attainments. I painted my
face, and to make myself as pitiable as possible I made a good scar and fixed one side of my lip in
a twist by the aid of a small slip of flesh-coloured plaster. Then with a red head of hair, and an
appropriate dress, I took my station in the business part of the city, ostensibly as a match-seller
but really as a beggar. For seven hours I plied my trade, and when I returned home in the evening I
found to my surprise that I had received no less than 26s. 4d.

"I wrote my articles and thought little more of the matter until, some time later, I backed a bill
for a friend and had a writ served upon me for £25. I was at my wit's end where to get the money,
but a sudden idea came to me. I begged a fortnight's grace from the creditor, asked for a holiday
from my employers, and spent the time in begging in the City under my disguise. In ten days I had
the money and had paid the debt.

"Well, you can imagine how hard it was to settle down to arduous work at £2 a week when I knew that
I could earn as much in a day by smearing my face with a little paint, laying my cap on the ground,
and sitting still. It was a long fight between my pride and the money, but the dollars won at last,
and I threw up reporting and sat day after day in the corner which I had first chosen, inspiring
pity by my ghastly face and filling my pockets with coppers. Only one man knew my secret. He was the
keeper of a low den in which I used to lodge in Swandam Lane, where I could every morning emerge as
a squalid beggar and in the evenings transform myself into a well-dressed man about town. This
fellow, a Lascar, was well paid by me for his rooms, so that I knew that my secret was safe in his
possession.