71 Moor End
"Some ten or twelve, but none which present any feature of interest. They are important, you
understand, without being interesting. Indeed, I have found that it is usually in unimportant
matters that there is a field for the observation, and for the quick analysis of cause and effect
which gives the charm to an investigation. The larger crimes are apt to be the simpler, for the
bigger the crime the more obvious, as a rule, is the motive. In these cases, save for one rather
intricate matter which has been referred to me from Marseilles, there is nothing which presents any
features of interest. It is possible, however, that I may have something better before very many
minutes are over, for this is one of my clients, or I am much mistaken."

He had risen from his chair and was standing between the parted blinds gazing down into the dull
neutral-tinted London street. Looking over his shoulder, I saw that on the pavement opposite there
stood a large woman with a heavy fur boa round her neck, and a large curling red feather in a broad-
brimmed hat which was tilted in a coquettish Duchess of Devonshire fashion over her ear. From under
this great panoply she peeped up in a nervous, hesitating fashion at our windows, while her body
oscillated backward and forward, and her fingers fidgeted with her glove buttons. Suddenly, with a
plunge, as of the swimmer who leaves the bank, she hurried across the road, and we heard the sharp
clang of the bell.

"I have seen those symptoms before," said Holmes, throwing his cigarette into the fire. "Oscillation
upon the pavement always means an affaire de coeur. She would like advice, but is not sure that the
matter is not too delicate for communication. And yet even here we may discriminate. When a woman
has been seriously wronged by a man she no longer oscillates, and the usual symptom is a broken bell
wire. Here we may take it that there is a love matter, but that the maiden is not so much angry as
perplexed, or grieved. But here she comes in person to resolve our doubts."

As he spoke there was a tap at the door, and the boy in buttons entered to announce Miss Mary
Sutherland, while the lady herself loomed behind his small black figure like a full-sailed merchant-
man behind a tiny pilot boat. Sherlock Holmes welcomed her with the easy courtesy for which he was
remarkable, and, having closed the door and bowed her into an armchair, he looked her over in the
minute and yet abstracted fashion which was peculiar to him.