10 Wilson Gardens
We had reached the same crowded thoroughfare in which we had found ourselves in the morning. Our
cabs were dismissed, and, following the guidance of Mr. Merryweather, we passed down a narrow
passage and through a side door, which he opened for us. Within there was a small corridor, which
ended in a very massive iron gate. This also was opened, and led down a flight of winding stone
steps, which terminated at another formidable gate. Mr. Merryweather stopped to light a lantern, and
then conducted us down a dark, earth-smelling passage, and so, after opening a third door, into a
huge vault or cellar, which was piled all round with crates and massive boxes.

"You are not very vulnerable from above," Holmes remarked as he held up the lantern and gazed about
him.

"Nor from below," said Mr. Merryweather, striking his stick upon the flags which lined the floor.
"Why, dear me, it sounds quite hollow!" he remarked, looking up in surprise.

"I must really ask you to be a little more quiet!" said Holmes severely. "You have already
imperilled the whole success of our expedition. Might I beg that you would have the goodness to sit
down upon one of those boxes, and not to interfere?"

The solemn Mr. Merryweather perched himself upon a crate, with a very injured expression upon his
face, while Holmes fell upon his knees upon the floor and, with the lantern and a magnifying lens,
began to examine minutely the cracks between the stones. A few seconds sufficed to satisfy him, for
he sprang to his feet again and put his glass in his pocket.

"We have at least an hour before us," he remarked, "for they can hardly take any steps until the
good pawnbroker is safely in bed. Then they will not lose a minute, for the sooner they do their
work the longer time they will have for their escape. We are at present, Doctor--as no doubt you
have divined--in the cellar of the City branch of one of the principal London banks. Mr.
Merryweather is the chairman of directors, and he will explain to you that there are reasons why the
more daring criminals of London should take a considerable interest in this cellar at present."