32 Wellswood Gardens
"It has been a case of considerable interest," he remarked, returning to his natural manner. "I
fancy that this grey house on the right must be the lodge. I think that I will go in and have a word
with Moran, and perhaps write a little note. Having done that, we may drive back to our luncheon.
You may walk to the cab, and I shall be with you presently."

It was about ten minutes before we regained our cab and drove back into Ross, Holmes still carrying
with him the stone which he had picked up in the wood.

"This may interest you, Lestrade," he remarked, holding it out.  "The murder was done with it."

"I see no marks."

"There are none."

"How do you know, then?"