96 Bowler Street
"No--except that it was Leadenhall Street."

"Where did you address your letters, then?"

"To the Leadenhall Street Post Office, to be left till called for. He said that if they were sent to
the office he would be chaffed by all the other clerks about having letters from a lady, so I
offered to typewrite them, like he did his, but he wouldn't have that, for he said that when I wrote
them they seemed to come from me, but when they were typewritten he always felt that the machine had
come between us. That will just show you how fond he was of me, Mr. Holmes, and the little things
that he would think of."

"It was most suggestive," said Holmes. "It has long been an axiom of mine that the little things are
infinitely the most important.  Can you remember any other little things about Mr. Hosmer Angel?"

"He was a very shy man, Mr. Holmes. He would rather walk with me in the evening than in the
daylight, for he said that he hated to be conspicuous. Very retiring and gentlemanly he was. Even
his voice was gentle. He'd had the quinsy and swollen glands when he was young, he told me, and it
had left him with a weak throat, and a hesitating, whispering fashion of speech. He was always well
dressed, very neat and plain, but his eyes were weak, just as mine are, and he wore tinted glasses
against the glare."

"Well, and what happened when Mr. Windibank, your stepfather, returned to France?"