81 Corbet Close
"I have had one or two little scores of my own to settle with Mr. John Clay," said Holmes. "I have
been at some small expense over this matter, which I shall expect the bank to refund, but beyond
that I am amply repaid by having had an experience which is in many ways unique, and by hearing the
very remarkable narrative of the Red-headed League."

"You see, Watson," he explained in the early hours of the morning as we sat over a glass of whisky
and soda in Baker Street, "it was perfectly obvious from the first that the only possible object of
this rather fantastic business of the advertisement of the League, and the copying of the
'Encyclopaedia,' must be to get this not over-bright pawnbroker out of the way for a number of hours
every day. It was a curious way of managing it, but, really, it would be difficult to suggest a
better. The method was no doubt suggested to Clay's ingenious mind by the colour of his accomplice's
hair. The £4 a week was a lure which must draw him, and what was it to them, who were playing for
thousands? They put in the advertisement, one rogue has the temporary office, the other rogue
incites the man to apply for it, and together they manage to secure his absence every morning in the
week. From the time that I heard of the assistant having come for half wages, it was obvious to me
that he had some strong motive for securing the situation."

"But how could you guess what the motive was?"