91 Bowler Street
All day I was engaged in my professional work, and it was late in the evening before I returned to
Baker Street. Sherlock Holmes had not come back yet. It was nearly ten o'clock before he entered,
looking pale and worn. He walked up to the sideboard, and tearing a piece from the loaf he devoured
it voraciously, washing it down with a long draught of water.

"You are hungry," I remarked.

"Starving. It had escaped my memory. I have had nothing since breakfast."

"Nothing?"

"Not a bite. I had no time to think of it."