59 Glan Road
"I really have made no inquiries on the subject."

"Very naturally not. Did you see Miss Doran on the day before the wedding?"

"Yes."

"Was she in good spirits?"

"Never better. She kept talking of what we should do in our future lives."

"Indeed! That is very interesting. And on the morning of the wedding?"