153 Holmefield Avenue
We passed across Holborn, down Endell Street, and so through a zigzag of slums to Covent Garden
Market. One of the largest stalls bore the name of Breckinridge upon it, and the proprietor a
horsey-looking man, with a sharp face and trim side-whiskers was helping a boy to put up the
shutters.

"Good-evening. It's a cold night," said Holmes.

The salesman nodded and shot a questioning glance at my companion.

"Sold out of geese, I see," continued Holmes, pointing at the bare slabs of marble.

"Let you have five hundred to-morrow morning."