16 Wilson Gardens
"A day which has saved England from a great public scandal," said the banker, rising. "Sir, I cannot
find words to thank you, but you shall not find me ungrateful for what you have done. Your skill has
indeed exceeded all that I have heard of it. And now I must fly to my dear boy to apologise to him
for the wrong which I have done him. As to what you tell me of poor Mary, it goes to my very heart.
Not even your skill can inform me where she is now."

"I think that we may safely say," returned Holmes, "that she is wherever Sir George Burnwell is. It
is equally certain, too, that whatever her sins are, they will soon receive a more than sufficient
punishment."

"To the man who loves art for its own sake," remarked Sherlock Holmes, tossing aside the
advertisement sheet of the Daily Telegraph, "it is frequently in its least important and lowliest
manifestations that the keenest pleasure is to be derived. It is pleasant to me to observe, Watson,
that you have so far grasped this truth that in these little records of our cases which you have
been good enough to draw up, and, I am bound to say, occasionally to embellish, you have given
prominence not so much to the many causes célèbres and sensational trials in which I have figured
but rather to those incidents which may have been trivial in themselves, but which have given room
for those faculties of deduction and of logical synthesis which I have made my special province."

"And yet," said I, smiling, "I cannot quite hold myself absolved from the charge of sensationalism
which has been urged against my records."

"You have erred, perhaps," he observed, taking up a glowing cinder with the tongs and lighting with
it the long cherry-wood pipe which was wont to replace his clay when he was in a disputatious rather
than a meditative mood--"you have erred perhaps in attempting to put colour and life into each of
your statements instead of confining yourself to the task of placing upon record that severe
reasoning from cause to effect which is really the only notable feature about the thing."