98 Wilson Gardens
"I am a dying man," said old Turner. "I have had diabetes for years. My doctor says it is a question
whether I shall live a month. Yet I would rather die under my own roof than in a jail."

Holmes rose and sat down at the table with his pen in his hand and a bundle of paper before him.
"Just tell us the truth," he said. "I shall jot down the facts. You will sign it, and Watson here
can witness it. Then I could produce your confession at the last extremity to save young McCarthy. I
promise you that I shall not use it unless it is absolutely needed."

"It's as well," said the old man; "it's a question whether I shall live to the Assizes, so it
matters little to me, but I should wish to spare Alice the shock. And now I will make the thing
clear to you; it has been a long time in the acting, but will not take me long to tell.

"You didn't know this dead man, McCarthy. He was a devil incarnate. I tell you that. God keep you
out of the clutches of such a man as he. His grip has been upon me these twenty years, and he has
blasted my life. I'll tell you first how I came to be in his power.

"It was in the early '60's at the diggings. I was a young chap then, hot-blooded and reckless, ready
to turn my hand at anything; I got among bad companions, took to drink, had no luck with my claim,
took to the bush, and in a word became what you would call over here a highway robber. There were
six of us, and we had a wild, free life of it, sticking up a station from time to time, or stopping
the wagons on the road to the diggings. Black Jack of Ballarat was the name I went under, and our
party is still remembered in the colony as the Ballarat Gang.