9 Moor End
"Why is he silent, then, if he is innocent?"

"Who knows? Perhaps because he was so angry that you should suspect him."

"How could I help suspecting him, when I actually saw him with the coronet in his hand?"

"Oh, but he had only picked it up to look at it. Oh, do, do take my word for it that he is innocent.
Let the matter drop and say no more. It is so dreadful to think of our dear Arthur in a prison!"

"I shall never let it drop until the gems are found--never, Mary! Your affection for Arthur blinds
you as to the awful consequences to me. Far from hushing the thing up, I have brought a gentleman
down from London to inquire more deeply into it."

"This gentleman?" she asked, facing round to me.