68 Down Close
"'The firm does so when the security is good.' I answered.

"'It is absolutely essential to me,' said he, 'that I should have £50,000 at once. I could, of
course, borrow so trifling a sum ten times over from my friends, but I much prefer to make it a
matter of business and to carry out that business myself. In my position you can readily understand
that it is unwise to place one's self under obligations.'

"'For how long, may I ask, do you want this sum?' I asked.

"'Next Monday I have a large sum due to me, and I shall then most certainly repay what you advance,
with whatever interest you think it right to charge. But it is very essential to me that the money
should be paid at once.'

"'I should be happy to advance it without further parley from my own private purse,' said I, 'were
it not that the strain would be rather more than it could bear. If, on the other hand, I am to do it
in the name of the firm, then in justice to my partner I must insist that, even in your case, every
businesslike precaution should be taken.'

"'I should much prefer to have it so,' said he, raising up a square, black morocco case which he had
laid beside his chair.  'You have doubtless heard of the Beryl Coronet?'