149 Bowler Street
"At Reading I had to change not only my carriage but my station. However, I was in time for the last
train to Eyford, and I reached the little dim-lit station after eleven o'clock. I was the only
passenger who got out there, and there was no one upon the platform save a single sleepy porter with
a lantern. As I passed out through the wicket gate, however, I found my acquaintance of the morning
waiting in the shadow upon the other side. Without a word he grasped my arm and hurried me into a
carriage, the door of which was standing open. He drew up the windows on either side, tapped on the
wood-work, and away we went as fast as the horse could go."

"One horse?" interjected Holmes.

"Yes, only one."

"Did you observe the colour?"

"Yes, I saw it by the side-lights when I was stepping into the carriage. It was a chestnut."