17 Bowler Street
"It's all clear," he whispered. "Have you the chisel and the bags? Great Scott! Jump, Archie, jump,
and I'll swing for it!"

Sherlock Holmes had sprung out and seized the intruder by the collar. The other dived down the hole,
and I heard the sound of rending cloth as Jones clutched at his skirts. The light flashed upon the
barrel of a revolver, but Holmes' hunting crop came down on the man's wrist, and the pistol clinked
upon the stone floor.