46 Harvey Street
"I signed the paper as directed, and the lawyer took it away with him. The singular incident made,
as you may think, the deepest impression upon me, and I pondered over it and turned it every way in
my mind without being able to make anything of it. Yet I could not shake off the vague feeling of
dread which it left behind, though the sensation grew less keen as the weeks passed and nothing
happened to disturb the usual routine of our lives. I could see a change in my uncle, however. He
drank more than ever, and he was less inclined for any sort of society. Most of his time he would
spend in his room, with the door locked upon the inside, but sometimes he would emerge in a sort of
drunken frenzy and would burst out of the house and tear about the garden with a revolver in his
hand, screaming out that he was afraid of no man, and that he was not to be cooped up, like a sheep
in a pen, by man or devil. When these hot fits were over, however, he would rush tumultuously in at
the door and lock and bar it behind him, like a man who can brazen it out no longer against the
terror which lies at the roots of his soul. At such times I have seen his face, even on a cold day,
glisten with moisture, as though it were new raised from a basin.

"Well, to come to an end of the matter, Mr. Holmes, and not to abuse your patience, there came a
night when he made one of those drunken sallies from which he never came back. We found him, when we
went to search for him, face downward in a little green-scummed pool, which lay at the foot of the
garden. There was no sign of any violence, and the water was but two feet deep, so that the jury,
having regard to his known eccentricity, brought in a verdict of 'suicide.' But I, who knew how he
winced from the very thought of death, had much ado to persuade myself that he had gone out of his
way to meet it. The matter passed, however, and my father entered into possession of the estate, and
of some £14,000, which lay to his credit at the bank."

"One moment," Holmes interposed, "your statement is, I foresee, one of the most remarkable to which
I have ever listened. Let me have the date of the reception by your uncle of the letter, and the
date of his supposed suicide."

"The letter arrived on March 10, 1883. His death was seven weeks later, upon the night of May 2nd."

"Thank you. Pray proceed."