89 Badgers Dene
"Your boy, Arthur, went to bed after his interview with you but he slept badly on account of his
uneasiness about his club debts. In the middle of the night he heard a soft tread pass his door, so
he rose and, looking out, was surprised to see his cousin walking very stealthily along the passage
until she disappeared into your dressing-room. Petrified with astonishment, the lad slipped on some
clothes and waited there in the dark to see what would come of this strange affair. Presently she
emerged from the room again, and in the light of the passage-lamp your son saw that she carried the
precious coronet in her hands. She passed down the stairs, and he, thrilling with horror, ran along
and slipped behind the curtain near your door, whence he could see what passed in the hall beneath.
He saw her stealthily open the window, hand out the coronet to someone in the gloom, and then
closing it once more hurry back to her room, passing quite close to where he stood hid behind the
curtain.

"As long as she was on the scene he could not take any action without a horrible exposure of the
woman whom he loved. But the instant that she was gone he realised how crushing a misfortune this
would be for you, and how all-important it was to set it right. He rushed down, just as he was, in
his bare feet, opened the window, sprang out into the snow, and ran down the lane, where he could
see a dark figure in the moonlight. Sir George Burnwell tried to get away, but Arthur caught him,
and there was a struggle between them, your lad tugging at one side of the coronet, and his opponent
at the other. In the scuffle, your son struck Sir George and cut him over the eye. Then something
suddenly snapped, and your son, finding that he had the coronet in his hands, rushed back, closed
the window, ascended to your room, and had just observed that the coronet had been twisted in the
struggle and was endeavouring to straighten it when you appeared upon the scene."