73 Corbet Close
"He must guard himself, for he may find that there is someone more cunning than himself upon his
track. You must lock yourself up from him to-night. If he is violent, we shall take you away to your
aunt's at Harrow. Now, we must make the best use of our time, so kindly take us at once to the rooms
which we are to examine."

The building was of grey, lichen-blotched stone, with a high central portion and two curving wings,
like the claws of a crab, thrown out on each side. In one of these wings the windows were broken and
blocked with wooden boards, while the roof was partly caved in, a picture of ruin. The central
portion was in little better repair, but the right-hand block was comparatively modern, and the
blinds in the windows, with the blue smoke curling up from the chimneys, showed that this was where
the family resided. Some scaffolding had been erected against the end wall, and the stone-work had
been broken into, but there were no signs of any workmen at the moment of our visit. Holmes walked
slowly up and down the ill-trimmed lawn and examined with deep attention the outsides of the
windows.

"This, I take it, belongs to the room in which you used to sleep, the centre one to your sister's,
and the one next to the main building to Dr. Roylott's chamber?"

"Exactly so. But I am now sleeping in the middle one."

"Pending the alterations, as I understand. By the way, there does not seem to be any very pressing
need for repairs at that end wall."

"There were none. I believe that it was an excuse to move me from my room."